[DEV-0002] Added Income and Expenses
Able to input Annual Income and Expenses (as percentages of income) for
now so that you're able to configure how much expenses you'll be using
out of your income as a percentage. it is still in progress to change.
This is only placed as a placeholder for the eventual intended effect
for it to be calculated in statistics.